#2 RENAME package main to miner
